Tweak Holy City religion spread chance
Increased the spread probability by 25% (not by 25 percentage points!) overall in order to even out a change that [advc.140] had made to the distance calculation.

And added a special clause for civs that still have 0 cities with the Holy City religion: increase the spread probability a little if such civs have fewer cities than the (map-size-based - e.g. 5 on Standard-size maps) target number of cities, and decrease the probability if the city count exceeds the target. This should make the overall chance of a religion spreading to a civ less dependent on its city count. [advc.173]

Vaguely related CFC post (2nd quote box):
https://forums.civfanatics.com/threads/advanced-civ.614217/page-47#post-16095451