Tweak Great Work culture
700 * (iEra + 1)
instead of
800 * iEra

In order to help strategies that aim for a Culture victory prior to the Modern era (i.e. to make sure that such strategies remain a consideration despite various changes that hurt culture rates especially in the early game and midgame). And for mod-mods that might allow a Great Artist in the Ancient era (iEra=0). [advc.251]

The era progression is now also described in the action button hover text and pointed out in the Civilopedia strategy help text.