Typedef CvInfo base classes for easier use of CvXMLInfo
I don't want to derive all the CvInfo classes from CvXMLInfo so long as they don't make use of the addElements function - because CvXMLInfo adds 24 byte of memory per instance (for two std::vectors). I do want to make it easier for mod-mods to change a base class to CvXMLInfo. That can now be done in a single place in the header file. The references to the base class in the implementation files now use the base class typedef from the header. [advc.tag]