Tweaks to AI evaluation of tech trades
Increase the impact of the utility that the AI assigns to techs that it receives (or that it'll impart to another AI civ). Previously, at most one third of the overall trade value had been subtracted when a tech had low utility, now up to 42% can be subtracted. And can add up to 27.5% (instead of 25%) when utility is high. [advc.550g]

Reduce early-game discounts for weak civs (score of known techs, military power level) in tech trades. [advc.550a]

Both changes are intended to make "trade bait" less powerful on the highest difficulty levels. Should e.g. no longer be possible to trade Aesthetics (which the AI will rarely prioritize - because it's not useful early on) 1-for-1 for Alphabet or Mathematics -- unless the recipient is actually somewhat keen on Aesthetics.

Not going to accomplish a great deal I think; Aesthetics and Compass (through a Great Scientist) are still going to be powerful - and often crucial - early techs on Immortal and Deity. Hopefully, things will at least no longer be worse than in BtS.