Sync info about civs having been randomly chosen
Doesn't really help with the True Starts option - which builds on that information - because the info can't be synchronized during game setup; or at least I haven't found a way. But at least it gets synchronized after game setup now so that all players in a network game can see on the Settings tab which civs and leaders were randomly assigned.
Amends change [advc.190c].