Fix loading of bonus resource data from XML
Instead of the era numbers, random junk had been loaded. (Which still worked half of the time when starting in the Ancient era, so it took me a while to notice.)

Can now load XML tag pairs that consist of two enum values (represented as strings in XML). Without this functionality, my macro for enum-to-enum CvInfo maps wasn't really useful. [advc.enum]