Fitness eval for precipitation
A more complex algorithm than I had originally envisioned. A first version (not committed or saved) had calculated a weighted mean of the precipitation for the entire surroundings of a starting plot. The surroundings are too large for this, the mean values end up too close to the global mean. The new algorithm instead divides the surroundings of a starting plot into overlapping regions, each having the size of a city radius, and a weighted mean is calculated for each region. Then an overall precipitation value is calculated only from a representative subset of those per-region precipitation values. This seems to match the intuitive impression that a human player gets of a starting plots and its surroundings pretty well.