Decrease latitude fitness penalty in some special cases
Namely when a civ with a high latitude target (Russia, but mainly: Vikings) is considered for a starting plot with a lower latitude, and when a civ with a subtropical latitude target is considered for a plot with a higher latitude. This latter change is aimed at desert-loving civs. Starting plots with a lot of nearby desert are scarce; if they're found outside the subtropics, it has to be good enough. Better if they're not too far from the jungle (and the Vikings not too far from the tundra), but, even on Large maps, no plot is really far from any climate zone - the vertical distances are quite short.