Place all map scripts in PrivateMaps
My understanding is that having the modified BtS scripts in PublicMaps allows CivilizationIV.ini to remember those scripts as the most recently played script. Really not important. It is neat to have the non-BtS scripts in a separate place. However, what's much more important to me is that Fractal (and not MixedContinents) appears at the top of the Custom Game screen. All scripts from PrivateMaps get listed before those in PublicMaps, so ...
(MixedContinents is fine, but players shouldn't assume that it's the mod's standard map script.) [advc.165]