Adjust some per-civ climate variation targets
Taking into account that this value is really only used as a target for variation in precipitation near a starting tile, so variation in (real-world) temperatures should not factor into the target values. I guess they still do to some extent, but I didn't want to change the targets too much - to avoid having to rebalance the iBias values.