Keep exact count of turns at peace
Change [advc.130k] had added randomness to the counter at CvTeamAI, and it turns out that the AI at-peace counter was being used for the "sustained peace" trade route modifier. That modifier should not be random - and shouldn't involve CvTeamAI at all - so I'm adding a separate at-peace counter to CvTeam.