Decrease AnimalAttackProb (except on Settler difficulty)
Now 70% rather than 90% on most difficulty levels. Mainly to make the animals' behavior appear a little more natural, and to make it so that players who want the animal to attack may have to stalk it a little bit. I guess I'd prefer a probability closer to 50%, but I don't want players to worry too much about this change (and the whole animal thing is still pretty terrible either way). [advc.309]