AI: Don't try to deny captives when last city besieged
Normally, it's a smart move to delete workers that a war enemy is about to capture, but, when they'll be captured along with the only remaining city, then there is no point in denying the captor anything. And there's always a small chance that the attack won't happen after all. [advc.010]