AI tech eval: count cities about to be founded
Much of the evaluation gets scaled by the current number of cities. I think, for the most part, this will work a little better in the early game if an extra city is counted when there's a settler in production or headed toward a city site. Should in particular help with the evaluation of religion techs (to be largely rewritten in another commit). [advc.131]