Overhaul AI eval of religion techs
[change id advc.171]

For the three early religions: Reduce the impact of randomness, turn number and city count and instead consider happiness needed in cities (i.e. in the capital) and the need for border expansion (in the 2nd city, though the capital will also help). As a result, an AI civ will usually go for a religion tech around the time that it founds a 2nd city. On average, this seems to be around the same time as in K-Mod (which founds religions a good deal later than BtS).

I've increased the impact of the urgency created by one or two of the early religion having already been taken. This should make it harder for a human player to found all three.

Flavor is still a big factor, and favorite religion arguably a bit more important than in K-Mod. (Seems to have had fairly little impact.)

For the later religions, I've mostly kept the K-Mod logic. Now trying harder to avoid founding a non-favorite religion, especially: not to be motivated by the favorite religion if getting there will likely result in founding another religion first. This is mainly relevant for the Abrahamic religions I think.

(I'll move the code into a subroutine in a separate commit.)