Fractal maps: choose smarter internal dimensions
The height maps generated by the Fractal algo normally use a 2:1 aspect ratio. Ought to be better to use 1:1 when the map dimensions are closer to that ratio. Fewer rounding errors that way I guess, if nothing else. (Since the Fractal dimensions have to be powers of 2, 2:1 and 1:1 are the only reasonable aspect ratios.) [advc.137]