RectLayout module, scale functions
A framework is taking shape for making this screen scalable:

RectLayout class for holding all the coordinates and dimensions of widgets or groups of widgets. With a constructor that allows specifying the coordinates as offsets from an enclosing "parent" widget. Also some parameters for center, left and right alignment within the parent. This should get rid of lots of simple but cumbersome calculations in CvMainInterface.

The rectangle data gets put in a global dictionary in CvMainInterface. This avoids typing 'self' all the time, and some other modules, e.g. PLE, will indeed need to access some of the data.

Wrapper functions around CyGInterfaceScreen that accept RectLayout objects (and with additional default parameters for convenience).

Global int -> int functions added that all scalable constants should pass through. One for horizontal side lengths of widgets, one for vertical side lengths, one for horizontal spaces in between widgets and one for vertical spaces. The idea is that spaces should arguably be less affected (if at all) by the screen resolution than widget dimensions. The (global) scale factors are still 1 at this point; will later want to set them based on the screen resolution.
The side lengths constants should be quite few in the end - most of the widgets should set their size and position based on other widgets instead.

So far, I'm only using this framework for the FoV slider and a bunch of panels. Next step will be to handle all widgets whose positional data I've already changed in the previous commits. Then go on with the remaining (half?) of the widgets that still have their absolute BtS positions.