Adjust HUD layout to avoid overlap w/ event log
[change id advc.092]

The default position of the event log (a.k.a. turn log) seems to be hardcoded in the EXE. The enlarged HUD had intersected with that. Will probably still be the case on screens higher than 1200 pixels (can't test that) - and in any case when the espionage slider is shown (i.e. when not hidden through change advc.120c).