Some tweaks to camera behavior
Through GlobalDefines that I had not been aware of previously. [advc.004m]

Increased camera scroll speed, i.e. faster movement parallel to the ground when pressing the arrow keys or when moving the mouse cursor to an edge of the screen. Related CFC thread:
https://forums.civfanatics.com/threads/trying-to-edit-scrollspeed-but-no-luck-any-idéas.674976

Camera distance on the city screen gets (slightly) adjusted to the field-of-view value. Can avoid issues like in this thread
https://forums.civfanatics.com/threads/city-map-zoom.676680/
that way.

Slightly steeper camera angle (more top-down). And slightly bigger field of view by default. Because the flatter camera angle that tends to result from a bigger field of view (player likely to zoom in a little farther) gets evened out by the direct change to the camera angle - and because the impact of the field of view on the size ratio between plots and plot indicators (balloons) is no longer a concern.