Make plot indicator runtime patch more robust
[change id advc.092b]

By falling back on a search for the appropriate region of the Civ4BeyondSword.exe code section when an instruction isn't found where the patch expects it. In case that e.g. localized editions of the game merely require a small offset to virtual addresses to be patched successfully. But I only have access to the MULTI5 version, so I don't know if this will accomplish anything.

And memory protections of the OS are now restored after the patch. They're, in principle, helpful for catching bugs, and turning the protections back on may also reduce the risk of being mistaken for a virus at runtime. (As for static virus checks, I've uploaded the DLL to VirusTotal.com now; no problems reported there. Apparently, just having VirtualProtect calls in the code does not raise suspicion.)

DLL updated (Assert build)