Darken research bar color
I had already darkened it a bit as part of some misc. color changes (advc.002i). Now that the research bar is much longer, I felt that it needed to be darkened further to make it less distracting. Perhaps it would really have to be shorter - but then the frame-like decorations on top will, again, not make sense. [advc.092]