Usability improvements for Sea Patrol
[change id advc.004k]

Allow Sea Patrol only when there is actually an adjacent improvement to guard. This way, the command button doesn't clutter the screen unnecessarily.

Highlight guarded tiles in blue when a unit on Sea Patrol is selected. (Previously, the player could only tell from the absence of the Sea Patrol button when a unit was on Sea Patrol.)

Show slow combat even if Quick Attack (but not Quick Defense) is enabled. To err on the side of making the player aware of this unexpected attack.