XML tag for per-tech adjustment to Barbarian diffusion
Using this to slow down the diffusion of Sailing (once the Barbarians have Fishing) so that Barbarian Galleys get delayed a little, and to speed up the diffusion of Horseback Riding (once the Barbarians have Animal Husbandry). This latter change replaces an earlier change that had sped up diffusion for techs with a high Military flavor value. [advc.301]