AI considers HR effect when canceling luxury trades
To address (at least one variety of) rare problems with AI civs canceling resource trades and immediately agreeing to a new trade with the exact same conditions. In the game state I had encountered, the AI was gaining 5 happiness from resources through 3 local luxuries, imported Spice and Hereditary Rule (with the AdvCiv effect of +25% happiness from luxuries). When making the deal, HR was not - and is still not - taken into account; when deciding to cancel the deal, the AI had assumed that it would lose only 1 happiness, not 2. This last part isn't entirely fixed - I'd rather not get into the weeds of anticipating how the HR effect gets rounded in each city -, but the issue should at least come up (even) more rarely now. [advc.912c]