Fix a sign error in AI-human war trades check
This seems to have existed in the mod since its initial release and caused the AI to treat humans as having nothing to trade when deciding whether to respond with "too much on our hands" - meaning that the AI had only offered to declare war when about to start war preparations of its own accord.

This commit also makes the AI a bit more imaginative wrt. trade items that the human player might be able to procure in the near future.