Fix production modifiers in calc. of overflow limit
Had been applying the production modifier of the next queued order (no modifier if none) instead of the modifiers that apply to the oder that was just completed. This issue was introduced by changes [advc.064b, advc.064d]. Rather minor because this only concerns the threshold at which production overflow gets converted into gold - most of the time, there isn't this much production overflow.