No speed adjustment for the production overflow cap
Now based (apart from the city's production rate) on the food that the city would need for growth on Normal speed and Noble difficulty - rather than the actual food needed for growth, which is adjusted to game speed and player handicap. Just a minor tweak to make conversion into gold a little less common on Quick speed and less common for the AI on the high difficulty levels. [advc.064b]

Related CFC post:
https://forums.civfanatics.com/threads/multiple-units-per-turn-unit-overflow-gold-wonder-fail-gold.683156/