Border spread at edge of Fog of War can trigger meeting
This change concerns a rare special case: Borders of an unmet rival spreading onto a fogged tile that is adjacent to a visible tile. In that case, the game will show the borders on the (whole) fogged tile (I guess so that the borders can then bleed onto the adjacent visible tile as they normally do), but no meeting occurs. I found this confusing, so a revealed tile (fully revealed but fogged) will now trigger a meeting if its owner becomes known. [advc.071b]