Macro for referring to worst enemy as "he" or "she"
To demonstrate how this can be achieved. Will require extra code for "He" and "She" (capitalized) and for the other grammatical cases, again with and without capitalization. Not great. Would be far easier to do if the relevant CvDiploParameters::setDiploComment call were in the DLL (as is the case for e.g. STOP_TRADING diplo comments; see CvPlayerAI:cpp); but apparently that call happens in the EXE in the case of WORST_ENEMY and WORST_ENEMY_TRADING. Additional translator macros are still less effort than reverse engineering and hacking. [advc.005h]