Espionage cost breakdown warns when civic/religion not practised
When asked to choose a civic or religion for the change-civics or convert-religion espionage mission, the mission cost breakdown says that the cost is increased when the spy owner does not practise the civic or religion. It doesn't say by how much exactly because the calculation is rather complicated and the intention is just to make the spy owner aware that their own civics/ religion affect the cost and, more importantly, that the ability to choose civics and religions other than their own is a deliberate change in AdvCiv and not a bug. [advc.132]

Related CFC post:
https://forums.civfanatics.com/threads/advanced-civ.614217/post-16695294