Allow checking BUG config (.ini files) prior to BUG init
I don't think this change will go on the main branch. I thought I might want it for a font size setting on the BUG menu, but it's all a little heavy-handed and ultimately not all that convenient to use. Certainly don't need this just for Sevopedia on the opening menu. (Sevopedia doesn't need to be optional at all, for all I care. It's a pretty clear improvement over the original Pedia.) [advc.009]

That being said, accessing the BUG config files without using the BUG engine is not terribly challenging to do - if it's really needed. I'm using configobj.py, which is a third-party library that comes with BUG.

Shortcomings:

The result of this early read of the BUG config is discarded once the BUG engine read those files for a second time. To reduce the inefficiency of this, I'm in fact only reading two of the config files - the ones I deem relevant for early access.

Early access is only available from the DLL, which needs to provide the path of the config folder - as BugPath.py isn't initialized yet - and I don't think the BugPath methods for figuring out the path would work this early on. The DLL gets the path from a memory location that seems to be hardcoded in the EXE and that I spotted by chance at some point (while looking for places where the EXE stores settings from CivilizationIV.ini). It's unclear if the path will really be stored at that location at all times and if it's really always the path that corresponds to the BUG config path.

DLL calls for early options access need to indicate that it's going to be an early access through an extra parameter (so that the path can get passed along), and rather than the options id (e.g. "MainInterface"), the section id (e.g. "Main") needs to be provided in the call arguments.