Add high resolution waiting cursor
Cherry-picked from Kek-Mod (Dawn of Civilization ultimately?):
https://github.com/DarkLunaPhantom/Kek-Mod/commit/72277e7b5078a85736d6be46ed5f4aa98ab28a83

When this was first added to Kek-Mod, I felt that the resolution is too high. I still think so, but the low resolution of the original cursor is even more jarring to me now, and I intend to increase the resolution of the yield icons too - and perhaps, in time, of some other graphics. [kekm.40]