Adjust position of the city billboard and its text nodes
I've only touched some "Translation" vectors (using NifSkope 1.1.3). Moved the whole thing up a tiny bit and a little to the left through the Master node (the text anchors had to be moved separately by the same amount). And centered the city name and production string more properly. These tweaks are probably for the better in any case, but all the more so as I'm about to scale up the city billboards through fader settings. [advc.002b]

Commit e0ea5e7adb1d3d7599bca350c4bb7d5609a57c58 (which probably won't be merged with the master branch) makes similar changes, but it applies them to an enlarged version of CityBar.nif.