AI war-declaring moves anticipate loss of Open Borders
That is, the pathfinder no longer assumes that routes of a war target can be used when moving toward one of the target's cities. Impact on performance not measured, hopefully negligible given how complex the actual movementCost function is; checking the DoW check beforehand shouldn't move the dial.

This change should make it much less likely that the path changes upon declaring war. In particular, the (already rare) phenomenon of an AI war declaration that is not immediately followed by units crossing the border will now, hopefully, be exceedingly rare. [advc.001t] (Also attaching that change id to a similar change previously covered by advc.082.)