Cancel random event popup if trigger conditions no longer met
Random events trigger at the end of a player's turn. If that player is human, the choice how to handle the event won't happen until the start of the player's next turn (via a popup). By then, the event may no longer make sense. So I'm now re-checking the trigger conditions before showing the popup. [advc.001]

It might be that this approach won't work for certain events, e.g. if an event could only trigger, say, in a leap year. My best bet is to investigate any canceled events for a while; so I've left an assertion to that end in place. Getting popups to launch during AI Auto Play seems challenging, so this will really have to be tested with a human player in charge.

This CFC thread made me aware of the delay between event trigger and popup:
https://forums.civfanatics.com/threads/forced-to-provide-a-resource-brothers-in-faith-at-war.693778/