AI eval. of strategic resources values a variety of units more
Specifically a variety of unit combat types. I had noticed that, once the AI has access to Chariots, it has little interest in Copper. This is not so wrong (Chariot should be fine until Iron gets revealed) and mostly has to do with Axeman having "Attack" as its default unit AI type (and thus being evaluated only for that role and not as a city attacker). But just having a good Melee attacker in addition to a Mounted attacker should be valuable. Still, this twea doesn't make much of a difference on the bottom line. [advc.036]

Related CFC post (above the 2nd quote box):
https://forums.civfanatics.com/threads/advanced-civ.614217/post-16754529