Cleaner handling of discover chance in AI improv. eval.
This way, existing Mines are treated as having a baseline value of 0, the same as any/most(?) other existing improvement that the AI considers to replace, and therefore should not be affected by the "improve better tiles first" logic. I'm not enthusiastic about the whole approach of evaluating differences from the current improvement (rather than the difference from a hypothetical unimproved tile - or more ambitiously the city's total yields), but it should at least be a little more consistent this way. [advc.121]