Apply AI improvement personality weights additively
My previous method was still flawed and may have contributed to oscillation problems. Now the AI takes the personality weight times a scale factor and adds the result to the comparison value of the improvement yields (comparing one improvement under consideration with the one currently present in the tile). As the scale factor, I use the smallest of the yield priority values, so that a 100% personality weight will usually mean treating an improvement as if it granted one extra commerce. The actual weights in XML are a good deal smaller than 100% (currently either 20% or 30%). [advc.131, advc.005a]