Add example usage of CivicWeightModifers to LeaderHead XML iWeightModifier represents a percentage value that each Leader would prefer the CivicType defined in their AI personality (beyond the favorite civic itself).
The AI's value calculation of which Civics he or she will adopt is directly multiplied by this weight modifier. See CvPlayerAI.cpp for the math involved.

I suggest to provide small positive values like 25 or 33 for Civics the Leader isn't particularly interested in but that helps to provide AI personality flavor. Example: Celestia with a Slavery preference based on alternative realities of her character.

Whereas the player could set values of 50 or 66 for Civics that this AI is both interested in using based on history and their personality. Example: Sefina Zono, the flying ponybot sourced from Ace Combat series, with Nationhood preference for extensive drafting into military service.

Finally a value of 75 or especially 100 would indicate that the Civic(s) specified for the Leader's personality weights should be used almost exclusively if possible. Example: King Sombra with a preference against the Emancipation Civic.