Create new branch - AI LeaderHead Favorite Tech Weights
AI players can have their own specific favorite technologies to prioritize.

Examples: Princess Celestia favoring Monarchy because of her long rule of Rquesteia and Education because of the university she sponsors.
Twilight Sparkle favoring Literature because of her love of books.
Rusa Imperio favoring Gunpowder and Rifling because of her origins in a war-torn universe...

As it is currently, each AI player seems to select the same set of technologies in each turn whenever they finish research as observed by spies or in Debug Mode. The idea is to create specific suggestions for each AI so that espionage can be more worthwhile and that everypony is not doing the same things all the time.