Add basic code structure for AI Personality Flavor based on Tech Weights
Enables the use of TechWeightModifiers to allow AI players to prefer certain technologies in their XML definitions. This helps AI players to not select the same Techs for everypony in the game.

A Princess Celestia player should favor some technologies differently than a King Sombra player, but the establish Flavors don't offer enough specificity for this purpose.