Alter Favorite Technology Weighting Calculations
Ensure that AI players do favor their preferred technologies but do not make them beholden to those preferences, so that the player advisors can provide guidance that will be listened to and taken into account.

Lessens the effect of the Favorite Technology on the calculation of the TechValue for eTech.