Add game code for handling Temporary Promotions
Enables the new XML tag "iExpireTurns", an integer value controlling how long a specific Promotion will last on a Unit.

For iExpireTurns, the granted Promotion will continue to operate until the counter on the Unit reaches 0, where it will mark the Promotion for removal.

If iExpireTurns is -1, the Promotion will never be considered a temporary ability and will instead work as in standard BtS (i.e. once earned, a Promotion stays on the Unit until death).