Add example usage of iExpireTurns in Promotions Info XML
Usage:
iExpireTurns=3
After 3 turns from when the Promotion was earned, the Promotion will expire and be removed from the Unit.

This would be useful for temporary enchantments or curses in the form of Promotions.

Example:
PROMOTION_HASTE
iExpireTurns = 1
A Unit with Haste would get several extra Moves for a limited period of time, exactly 1 full Turn since when a spellweaver cast the enchantment upon the Unit.