AI Fights Enemy Civics or Religions in Diplomacy
Leaders with a Religion or Civic they oppose won't consider an offer from the player (or other AI player) that proposes the adoption of the Enemy Civic/Religion.

It could be useful to have an option to check if a Leader is intolerant of a/any Religion or not, as shown in Fall From Heaven 2. The devised code currently uses isHasReligion, which evaluates if the Religion is present in any City or not. Tolerant AI players could instead check for getStateReligion being equal to the enemy Religion they oppose, allowing for tolerant players to trade a Religion their trade partner is not actively supporting.