Add Example of First Impression Diplomacy Modifiers
Create new XML tag "FirstImpressionChanges" and attitude tag "iFirstImpressionAttitudeChange" to control how a Leader sees the Other Leader named in the tags.

Example:

If the Examining Leader's Type = LEADER_CELESTIA
and FirstImpressionChanges contains the tag Type = LEADER_LUNA with an AttitudeChange of 1

Then Princess Celestia's First Impressions modifier at the start of a game will be increased by additional 1 from whatever it might have been randomly calculated to.
If Celestia's First Impression was calculated as 2 randomly, it will then be increased to 3 for this game only.