Create new branch - Leader Specific Colony Vassals
Each Leader can prefer to put in charge of their Colonies a specific Leader, for example Princess Celestia might put Twilight Sparkle on the throne of her Equestrian colony.

Alternatively, Princess Twilight Sparkle would place Luster Dawn in charge. It depends on if the colony Master has an existing lore relationship with the Leader in question (and the spot not being taken in-game), otherwise the decision would be randomized.