Create new branch - Armageddon Counter Import from Fall From Heaven II
The Armageddon Counter from FfH II would be very interesting to have in MLP Civilization is Magic, as it brings depth to the strategy of combat/unit compositions.

In this way I can simulate the magical effects of Promotions and the power shown in fan-made worlds like the MLP Loops project and the Filly Fantasy VI game mod.

The Armageddon Counter would also enhance the lore of my own constructed world where the Light Goddess Solais a Tuath and the Dark Goddess Aequilibrius ze Aternum battle endlessly for supremacy, as it could be utilized through Events and combat to offer certain Leaders/Civilizations reasons to increase/decrease it.

E.g. Good/Evil Diplomacy, Unit Resurrections, Event Potency/Frequency Modification, and so forth.