Add Area Counter for Tracking Number of Mystical Tiles
Mystical tiles represent those tiles where the local area has transformed into a magical location, such as Heaven/Hell, or anywhere that an Area has undergone a shift in spiritual potential from Citizens being born or displaced.

The idea here is to mimic the untamed Area that is the Everfree Forest from the MLP show, somewhat, in game terms and provide for certain players to favor transforming the local Area for their own purposes.