Create new branch - AI Survivor In-Game Stat Tracking
Viewing in-game stats like # of Kills or # of Victories over a whole season of game play would be exciting for AI role-playing.

This would also enable development of interesting features like special diplomatic memories persisting over multiple games and AI player strategy alteration based on what has worked for them in the past.

For example if an AI player King Sombra knows that a Domination or Conquest victory has worked for him in the past, his AI personality could alter the strategy at the beginning of a game to account for this and thus pursue the victory more often (or less often as the case might be).