Create the basic structure of the CvEquipmentInfo class
A new CvInfo class related to Units, the Equipment class offers combat-related abilities and Promotions for as long as the Unit possesses the Equipment piece in question.

For example, the Alicorn Amulet boosts a Unit's combat power, but as long as it is worn by a pony it also grants a Berserker-type Promotion (and as soon as the Equipment piece is lost the Promotion is automatically removed from the Unit).