Add code for Leader Specific Team Size Attitude Change values
A positive value for iTeamSizeAttitudeChange reduces the penalty for a team size difference between the human player and the AI. Essentially, the AI will be less cautious about the team's power and thus unconcerned during diplomacy.

For example, Twilight Sparkle could have a positive value here and be more willing to work with a powerful human player.

A negative value for iTeamSizeAttitudeChange increases the penalty further, ensuring that an AI will be more cautious about the team's power disparity and more distrustful during diplomacy.

For example, King Sombra could have a negative value here and be more wary of a powerful human player challenging his own dominance.