Add DamageTypeResistPoints to game information alongside DamageTypeCombatPoints
These points are used in an upcoming function to give units access to magical abilities via weighted randomization. Each unit will get one ability based on which magical power has the most "points". The values can increase according to the terrain of a city, other units with the same power on the plot, the type of civilization or leader that someone is playing as, etc.

Also start writing the magical strength calculation code to use these values in unit-on-unit combat. Also clean up existing code as well.