Add ReputationScore class based on PlayerHistory class from AdvCiv
This class will be used to store the reputation score of an opposing player, the score effecting how this AI player views the opposing player.

For instance a player with a high value for Kindness may find that deals are easier to make with this AI player. However a player with a negative value for Loyalty may find themselves being considered as a target or possibly having a third party being paid to attack them.