Add initial C+++ code for Shared Relative Progress Cooperation values to be used with AI leaders
This is a more dynamic evaluation of two players struggling at the bottom of the score rankings and allows for AI leaders to be more or less cooperative with their fellow players based on the standings.

The AttitudeThreshold value determines how open an AI is to cooperating with those at a similar ranking level to itself.
An Annoyed leader is defined as more isolationist than a Friendly one under this system.

The AttitudeChange, AttitudeDivisor, and AttitudeChangeLimit values determines by how much the AI values cooperation with anyone at a similar ranking level. The system also supports negative values that would make an AI dislike anyone struggling along with it, providing for more varied diplomatic behavior and AI personality.

Todo: Add display support (CvGameTextMgr) for the new diplomacy values and UWAI personality factor weight calculations.